src/app/loader/loader.component.ts
import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
import { Subscription } from 'rxjs/internal/Subscription';
import { LoaderService } from './shared/loader.service';

@Component({
    selector: 'app-loader',
    templateUrl: './loader.component.html',
    styleUrls: ['./loader.component.scss'],
    changeDetection: ChangeDetectionStrategy.OnPush
})
export class LoaderComponent implements OnInit {
    constructor(public loaderService: LoaderService) { }

    ngOnInit() {
    }
}